requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.4.0
asgiref==3.4.1
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
datasets==1.16.1
dill==0.3.4
fastapi==0.70.0
fastapi-utils==0.2.1
filelock==3.4.0
flatbuffers==2.0
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
h11==0.12.0
h5py==3.6.0
huggingface-hub==0.2.0
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
multidict==5.2.0
multiprocess==0.70.12.2
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyDeprecate==0.3.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytorch-lightning==1.5.4
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.8.0+cpu
torchmetrics==0.6.0
tqdm==4.62.3
transformers==4.12.5
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2
wrapt==1.13.3
xxhash==2.0.2
yarl==1.7.2
zipp==3.6.0
